requirements.txt
packaging==24.2
pillow==11.1.0
PyMuPDF==1.25.5
PyPDF2==3.0.1
pytesseract==0.3.13
